refactor: Group 도메인의 값/권한 검증 코드 policy로 이동

+++ src/main/java/com/process/clash/adapter/persistence/group/GroupMemberJpaRepository.java
@@ -14,7 +14,7 @@ public interface GroupMemberJpaRepository extends JpaRepository<GroupMemberJpaEn
 
     boolean existsByGroupIdAndUserId(Long groupId, Long userId);
 
-    long countByGroupId(Long groupId);
+    int countByGroupId(Long groupId);
 
     void deleteByGroupIdAndUserId(Long groupId, Long userId);
